+++ src/main/java/papers/friendsAgenda/service/PersonService.java
@@ -36,7 +36,7 @@ public class PersonService {
         return personRepository.save(person);
     }
 
-    public void eliminar(Long id) {
+    public void eliminar(Integer id) {
         personRepository.deleteById(id);
     }
 }
